добавил простенькую валидацию для post и put

+++ api/models/AbstractModel.php
@@ -6,6 +6,16 @@ abstract class AbstractModel {
 
 	public $id = null;
 
+	public static function validation ($array)
+	{
+		foreach ($array as $key => $value) {
+			if (!array_key_exists($key, static::$attributes)) {
+				throw new Exception ("Ошибка в написании");
+			}
+		}
+		return 'the attributes are valid';
+	}
+
 	public static function getOne($id)
 	{
 	    $db_data = new Database();
@@ -22,6 +32,7 @@ abstract class AbstractModel {
 
 	public static function create($post)
 	{
+		echo self::validation($post);
 		foreach ($post as $key => $value){
 		    $post[$key] = "'".$value."'";
 		}
@@ -33,12 +44,11 @@ abstract class AbstractModel {
 
 	public static function update($id, $changed_data)
 	{
+		echo self::validation($changed_data);
 		$db_data = new Database();
 		print_r($changed_data);
 		foreach ($changed_data as $key => $value) {
-			if (array_key_exists($key, static::$attributes) != false) {
-				$data = $db_data->query('UPDATE '.static::$table_name.' SET '.$key." = "."'".$value."'"." ".'WHERE id = '.$id)."<br>";
-			}
+			$data = $db_data->query('UPDATE '.static::$table_name.' SET '.$key." = "."'".$value."'"." ".'WHERE id = '.$id)."<br>";
 		}
     	return $data;
     }
@@ -46,8 +56,9 @@ abstract class AbstractModel {
 	public static function delete($id)
 	{
 		$db_data = new Database();
-
-		$result = $db_data->query('DELETE FROM Appointments WHERE patient_id = '.$id);
+		if (static::$table_name = 'Patients') {
+			$result = $db_data->query('DELETE FROM Appointments WHERE patient_id = '.$id);
+		}
 		$result = $db_data->query('DELETE FROM '.static::$table_name.' WHERE id = '.$id);
 
 		return $result;
